backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
numpy==2.1.1
Pillow==10.4.0
tensorflow==2.21.0
pytest==8.3.3
gunicorn==22.0.0
